Setting slmsk to be integer (#569)
* setting slmsk to be integer

* removing hofx units coeff YAML key

* adjusting references and adding tolerances

* updating references

* reverting #571 changes

* removing spaces from reference

+++ src/fv3jedi/FieldMetadata/FieldsMetadataDefault.h
@@ -546,7 +546,7 @@ namespace fv3jedi {
     md.longName = "slmsk";
     md.shortName = "slmsk";
     md.units = "none";
-    md.kind = "double";
+    md.kind = "integer";
     md.tracer = "false";
     md.horizontalStaggerLocation = "center";
     md.levels = "1";
